+++ stac_pydantic/catalog.py
@@ -2,8 +2,9 @@ from typing import List, Optional
 
 from pydantic import BaseModel, Field, root_validator
 
+from stac_pydantic.links import Link, Links
+
 from .extensions import Extensions
-from .shared import Link, Links
 from .version import STAC_VERSION
