+++ stac_pydantic/item.py
@@ -6,9 +6,11 @@ from geojson_pydantic.features import Feature, FeatureCollection
 from pydantic import BaseModel, Field, create_model, validator
 from pydantic.fields import FieldInfo
 
+from stac_pydantic.links import Links
+
 from .api.extensions.context import ContextExtension
 from .extensions import Extensions
-from .shared import Asset, BBox, Links, StacCommonMetadata
+from .shared import Asset, BBox, StacCommonMetadata
 from .utils import decompose_model
 from .version import STAC_VERSION
