+++ stac_pydantic/shared.py
@@ -1,7 +1,6 @@
 from datetime import datetime
 from enum import Enum, auto
-from typing import Any, Dict, Iterator, List, Optional, Tuple, Union
-from urllib.parse import urljoin
+from typing import List, Optional, Tuple, Union
 
 from pydantic import BaseModel, Extra, Field
 
@@ -66,28 +65,6 @@ class MimeTypes(str, Enum):
     text = "text/plain"
 
 
-class Relations(str, AutoValueEnum):
-    """
-    https://github.com/radiantearth/stac-spec/blob/v1.0.0-beta.1/collection-spec/collection-spec.md#relation-types
-    """
-
-    self = auto()
-    root = auto()
-    parent = auto()
-    child = auto()
-    item = auto()
-    license = auto()
-    derived_from = auto()
-    collection = auto()
-    alternate = auto()
-    previous = auto()
-    next = auto()
-    conformance = auto()
-    docs = auto()
-    tiles = auto()
-    search = auto()
-
-
 class AssetRoles(str, AutoValueEnum):
     """
     https://github.com/radiantearth/stac-spec/blob/v1.0.0-beta.1/extensions/asset/README.md
@@ -104,68 +81,6 @@ class ProviderRoles(str, AutoValueEnum):
     producer = auto()
     processor = auto()
     host = auto()
-
-
-class Link(BaseModel):
-    """
-    https://github.com/radiantearth/stac-spec/blob/v1.0.0-beta.1/collection-spec/collection-spec.md#link-object
-    """
-
-    href: str
-    rel: str
-    type: Optional[str]
-    title: Optional[str]
-    # Label extension
-    label: Optional[str] = Field(None, alias="label:assets")
-
-    class Config:
-        use_enum_values = True
-
-    def resolve(self, base_url: str):
-        """resolve a link to the given base URL"""
-        self.href = urljoin(base_url, self.href)
-
-
-class PaginationMethods(str, AutoValueEnum):
-    """
-    https://github.com/radiantearth/stac-api-spec/blob/master/api-spec.md#paging-extension
-    """
-
-    GET = auto()
-    POST = auto()
-
-
-class PaginationRelations(str, AutoValueEnum):
-    """
-    https://github.com/radiantearth/stac-api-spec/blob/master/api-spec.md#paging-extension
-    """
-
-    next = auto()
-    previous = auto()
-
-
-class PaginationLink(Link):
-    """
-    https://github.com/radiantearth/stac-api-spec/blob/master/api-spec.md#paging-extension
-    """
-
-    rel: PaginationRelations
-    method: PaginationMethods
-    body: Optional[Dict[Any, Any]]
-    merge: bool = False
-
-
-class Links(BaseModel):
-    __root__: List[Union[PaginationLink, Link]]
-
-    def resolve(self, base_url: str):
-        """resolve all links to the given base URL"""
-        for link in self:
-            link.resolve(base_url)
-
-    def __iter__(self) -> Iterator[Link]:
-        """iterate through links"""
-        return iter(self.__root__)
 
 
 class Provider(BaseModel):
